end of sec18

+++ client/src/app/messages/messages.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Message } from '../_models/message';
 import { Pagination } from '../_models/pagination';
+import { ConfirmService } from '../_services/confirm.service';
 import { MessageService } from '../_services/message.service';
 
 @Component({
@@ -21,7 +22,7 @@ export class MessagesComponent implements OnInit {
   // functional properties
   loading: boolean = false;
 
-  constructor(private messageService: MessageService) { }
+  constructor(private messageService: MessageService, private confirmService: ConfirmService) { }
 
   ngOnInit(): void {
     this.loadMessages();
@@ -43,11 +44,17 @@ export class MessagesComponent implements OnInit {
 
   // Delete message
   deleteMessage(id: number) {
-    this.messageService.deleteMessage(id).subscribe(
-      () => {
-        this.messages.splice(this.messages.findIndex(m => m.id === id), 1);
+
+    this.confirmService.confirm('Confirm to delete messge', 'This cannot be undone').subscribe(result => {
+      if (result) {
+        this.messageService.deleteMessage(id).subscribe(
+          () => {
+            this.messages.splice(this.messages.findIndex(m => m.id === id), 1);
+          }
+        );
       }
-    )
+    })
+
   }
 
   // Functional methods:
